Move types over into @types dir
