+++ @types/CacheManager.d.ts
@@ -5,11 +5,9 @@ export interface CacheManager {
     clear(key: string): void;
 }
 
-export abstract class AbstractCacheManager implements CacheManager {
-    constructor(identifier = "") {
-        this.identifier = identifier;
-    }
-    identifier = "";
+export declare abstract class AbstractCacheManager implements CacheManager {
+    constructor(identifier?: string);
+    identifier: string;
     abstract get<T = any>(key: string): T;
     abstract set(key: string, value: any): void;
     abstract clear(key: string): void;
