+++ @types/Helpers.d.ts
@@ -1,11 +1,11 @@
 import u from "umbrellajs";
-import { Color, Style, Scroll, Position } from "../utils";
-import { Step } from "./Step";
+import { AbstractStep } from "./Step";
+import { Color, Position, Scroll, Style } from "../src/utils";
 
 export type Helpers = {
     assert(condition: any, message: string): boolean;
     clamp(value: number, min: number, max?: number): number;
-    decorate(text: string, step: Step): string;
+    decorate(text: string, step: AbstractStep): string;
     getMaxZIndex(): number;
     Color: typeof Color;
     Style: typeof Style;
