+++ src/step/CardStep.ts
@@ -1,5 +1,5 @@
-import { Position } from "../utils/position";
 import { Element } from "umbrellajs";
+import { Position } from "../utils/position";
 import PopoverStep from "./PopoverStep";
 
 interface CardStepData {
