+++ src/utils/style.ts
@@ -1,5 +1,5 @@
 import u, { Element } from "umbrellajs";
-import { TourStyle } from "../types";
+import { TourStyle } from "../../@types";
 
 // eslint-disable-next-line @typescript-eslint/no-namespace
 export namespace Style {
