+++ tsconfig.json
@@ -20,6 +20,6 @@
     ],
     "include": [
         "src",
-        "src/types"
+        "@types"
     ]
 }
