--- types/decorator/MarkdownDecorator.d.ts
@@ -1,2 +1,0 @@
-import { ContentDecorator } from "./ContentDecorator";
-export declare const MarkdownDecorator: ContentDecorator;
