--- types/handler/ActionHandler.d.ts
@@ -1,3 +1,0 @@
-import { ActionHandler as ActionHandlerType, Tour, TourAction } from "../types";
-export type ActionHandlerFn = (event: Event, action: TourAction, context: Tour) => void;
-export default function ActionHandler(name: string, handlerFn: ActionHandlerFn): ActionHandlerType;
