--- types/types/index.d.ts
@@ -1,6 +1,0 @@
-export * from "./ActionHandler";
-export * from "./CacheManager";
-export * from "./ContentDecorator";
-export * from "./Helpers";
-export * from "./Step";
-export * from "./Tour";
