--- types/utils/assert.d.ts
@@ -1,1 +1,0 @@
-export declare function assert(condition: boolean | string | number | any, message: string): boolean;
