--- types/utils/clamp.d.ts
@@ -1,1 +1,0 @@
-export declare function clamp(value: number, min: number, max?: number): number;
