--- types/utils/dom.d.ts
@@ -1,1 +1,0 @@
-export declare function getDataContents<T>(data?: string, defaults?: any): T;
