--- types/utils/index.d.ts
@@ -1,7 +1,0 @@
-export * from "./assert";
-export * from "./clamp";
-export * from "./color";
-export * from "./style";
-export * from "./zindex";
-export * from "./scroll";
-export * from "./position";
