--- types/utils/zindex.d.ts
@@ -1,1 +1,0 @@
-export declare function getMaxZIndex(): number;
